🎨 Texture change

+++ source/graphics_FA.h
@@ -142,8 +142,8 @@ typedef struct
     unsigned middle;
 
     //NOTE(Stanisz13): Left top corner is (0, 0), bottom right corner is (screenW, screenH);
-    unsigned posX;
-    unsigned posY;
+    float posX;
+    float posY;
     
     int wheel; //NOTE(Stanisz13): sign depending on scroll direction
